DB: warmth_rating → numeric(3,1) for 0.5-step values
The slider now emits 1.5, 2.5, etc but the column was still integer, so
Supabase rejected the PATCH with "invalid input syntax for type integer".
Switched the column to numeric(3,1) (three digits, one decimal) which
covers 1.0-5.0 neatly, and kept the between-1-and-5 check constraint.

Run supabase-migration-warmth-halfstep.sql against the live DB.

+++ supabase-schema.sql
@@ -50,7 +50,7 @@ create table clothing_items (
 
   seasons text[] default '{}'::text[],
   occasions text[] default '{}'::text[],
-  warmth_rating integer default 3 check (warmth_rating between 1 and 5),
+  warmth_rating numeric(3, 1) default 3 check (warmth_rating between 1 and 5),
   rain_appropriate boolean default false,
   brand text,
